Replace parts of system headers with <_bits/...>.

+++ kernel/arch/sparc64/include/arch/mm/sun4v/tsb.h
@@ -71,7 +71,7 @@ typedef struct tsb_descr {
 struct as;
 struct pte;
 
-extern void tsb_invalidate(struct as *as, uintptr_t page, uint64_t pages);
+extern void tsb_invalidate(struct as *as, uintptr_t page, size_t pages);
 extern void itsb_pte_copy(struct pte *t);
 extern void dtsb_pte_copy(struct pte *t, bool ro);
